+++ packages/react/.storybook/main.js
@@ -1,8 +1,8 @@
 module.exports = {
   addons: [
-    "@storybook/addon-essentials",
-    "@storybook/addon-links",
-    "@whitespace/storybook-addon-html",
+    '@storybook/addon-essentials',
+    '@storybook/addon-links',
+    '@whitespace/storybook-addon-html',
   ],
-  stories: ["../src/**/*.stories.@(mdx|tsx)"],
+  stories: ['../src/**/*.stories.@(mdx|tsx)'],
 };