+++ packages/react/postcss.config.js
@@ -2,8 +2,8 @@ module.exports = {
   extract: true,
   minimize: true,
   plugins: [
-    require("postcss-import"),
-    require("postcss-preset-env"),
-    require("postcss-custom-media"),
+    require('postcss-import'),
+    require('postcss-preset-env'),
+    require('postcss-custom-media'),
   ],
 };