+++ packages/react/src/__mocks__/fileMock.ts
@@ -1,3 +1,3 @@
 // src/__mocks__/mock/fileMock.ts
 
-export default "test-file-stub";
+export default 'test-file-stub';