+++ packages/react/src/components/Button/Button.stories.tsx
@@ -1,13 +1,13 @@
 // Button.stories.js
 
-import React from "react";
-import { Story } from "@storybook/react";
-import Button, { ButtonProps } from "./Button";
-import mdx from "./Button.mdx";
+import React from 'react';
+import { Story } from '@storybook/react';
+import Button, { ButtonProps } from './Button';
+import mdx from './Button.mdx';
 
 export default {
-  title: "Components/Button",
-  args: { children: "Button" },
+  title: 'Components/Button',
+  args: { children: 'Button' },
   parameters: {
     docs: {
       page: mdx,