+++ packages/react/src/components/Button/Button.test.tsx
@@ -1,10 +1,10 @@
-import { render, screen } from "@testing-library/react";
-import React from "react";
+import { render, screen } from '@testing-library/react';
+import React from 'react';
 
-import { Button } from "..";
+import { Button } from '..';
 
-describe("<Button />", () => {
-  test(": renders", () => {
+describe('<Button />', () => {
+  test(': renders', () => {
     render(<Button>Button</Button>);
     const Text = screen.getByText(/Button/i);