+++ packages/styles/rollup.config.js
@@ -1,13 +1,13 @@
-import postcss from "rollup-plugin-postcss";
+import postcss from 'rollup-plugin-postcss';
 
-const postcssConfig = require("./postcss.config.js");
+const postcssConfig = require('./postcss.config.js');
 
 export default [
   {
-    input: "src/style.scss",
+    input: 'src/style.scss',
     output: {
-      file: "dist/style.css",
-      format: "es",
+      file: 'dist/style.css',
+      format: 'es',
     },
     plugins: [postcss(postcssConfig)],
   },