+++ packages/tokens/scripts/build.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-const StyleDictionaryPackage = require("style-dictionary");
+const StyleDictionaryPackage = require('style-dictionary');
 
 // HAVE THE STYLE DICTIONARY CONFIG DYNAMICALLY GENERATED
 
@@ -7,97 +7,97 @@ function getStyleDictionaryConfig(brand, platform) {
   return {
     source: [
       `src/brands/${brand}/*.json`,
-      "src/globals/**/*.json",
+      'src/globals/**/*.json',
       `src/platforms/${platform}/*.json`,
     ],
     platforms: {
       styles: {
         transforms: [
-          "attribute/cti",
-          "name/cti/kebab",
-          "time/seconds",
-          "content/icon",
-          "color/css",
+          'attribute/cti',
+          'name/cti/kebab',
+          'time/seconds',
+          'content/icon',
+          'color/css',
         ],
         buildPath: `dist/web/${brand}/`,
         files: [
           {
-            destination: "css/tokens.css",
-            format: "css/variables",
+            destination: 'css/tokens.css',
+            format: 'css/variables',
             options: { showFileHeader: false },
           },
           {
-            destination: "scss/tokens.scss",
-            format: "scss/variables",
+            destination: 'scss/tokens.scss',
+            format: 'scss/variables',
             options: { showFileHeader: false },
           },
           {
-            destination: "less/tokens.less",
-            format: "less/variables",
+            destination: 'less/tokens.less',
+            format: 'less/variables',
             options: { showFileHeader: false },
           },
         ],
       },
       js: {
-        transforms: ["attribute/cti", "name/cti/pascal", "color/hex"],
+        transforms: ['attribute/cti', 'name/cti/pascal', 'color/hex'],
         buildPath: `dist/web/${brand}/`,
         files: [
           {
-            destination: "js/module/index.js",
-            format: "javascript/module",
+            destination: 'js/module/index.js',
+            format: 'javascript/module',
             options: { showFileHeader: false },
           },
           {
-            destination: "js/object/index.js",
-            format: "javascript/object",
+            destination: 'js/object/index.js',
+            format: 'javascript/object',
             options: { showFileHeader: false },
           },
           {
-            destination: "js/umd/index.js",
-            format: "javascript/umd",
+            destination: 'js/umd/index.js',
+            format: 'javascript/umd',
             options: { showFileHeader: false },
           },
           {
-            destination: "js/es6/index.js",
-            format: "javascript/es6",
+            destination: 'js/es6/index.js',
+            format: 'javascript/es6',
             options: { showFileHeader: false },
           },
           {
-            destination: "json/tokens.json",
-            format: "json",
+            destination: 'json/tokens.json',
+            format: 'json',
             options: { showFileHeader: false },
           },
         ],
       },
       android: {
-        transformGroup: "android",
+        transformGroup: 'android',
         buildPath: `dist/android/${brand}/`,
         files: [
           {
-            destination: "tokens.colors.xml",
-            format: "android/colors",
+            destination: 'tokens.colors.xml',
+            format: 'android/colors',
           },
           {
-            destination: "tokens.dimens.xml",
-            format: "android/dimens",
+            destination: 'tokens.dimens.xml',
+            format: 'android/dimens',
           },
           {
-            destination: "tokens.font_dimens.xml",
-            format: "android/fontDimens",
+            destination: 'tokens.font_dimens.xml',
+            format: 'android/fontDimens',
           },
         ],
       },
       ios: {
-        transforms: ["attribute/cti", "name/cti/camel", "color/UIColorSwift"],
+        transforms: ['attribute/cti', 'name/cti/camel', 'color/UIColorSwift'],
         buildPath: `dist/ios/${brand}/`,
         files: [
           {
-            destination: "tokens.h",
-            format: "ios/macros",
+            destination: 'tokens.h',
+            format: 'ios/macros',
           },
           {
-            destination: "tokens.swift",
-            format: "ios-swift/class.swift",
+            destination: 'tokens.swift',
+            format: 'ios-swift/class.swift',
           },
         ],
       },
@@ -105,13 +105,13 @@ function getStyleDictionaryConfig(brand, platform) {
   };
 }
 
-console.log("Build started...");
+console.log('Build started...');
 
 // PROCESS THE DESIGN TOKENS FOR THE DIFFEREN BRANDS AND PLATFORMS
 
-["mono"].map(function (brand) {
-  ["styles", "js", "ios", "android"].map(function (platform) {
-    console.log("\n==============================================");
+['mono'].map(function (brand) {
+  ['styles', 'js', 'ios', 'android'].map(function (platform) {
+    console.log('\n==============================================');
     console.log(`\nProcessing: [${platform}] [${brand}]`);
 
     const StyleDictionary = StyleDictionaryPackage.extend(
@@ -120,9 +120,9 @@ console.log("Build started...");
 
     StyleDictionary.buildPlatform(platform);
 
-    console.log("\nEnd processing");
+    console.log('\nEnd processing');
   });
 });
 
-console.log("\n==============================================");
-console.log("\nBuild completed!");
+console.log('\n==============================================');
+console.log('\nBuild completed!');