+++ packages/styles/CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ### Features
 
-- renaming components -> styles ([41c6a21](https://github.com/richmccartney/mono-test/commit/41c6a2139617bf63ebeabc31e5302fafd11df84d))
+- renaming components -> styles
+  ([41c6a21](https://github.com/richmccartney/mono-test/commit/41c6a2139617bf63ebeabc31e5302fafd11df84d))
 
 # mono-test-components [1.1.0-next.1](https://github.com/richmccartney/mono-test/compare/mono-test-components@1.0.0...mono-test-components@1.1.0-next.1) (2020-11-17)
 
 ### Features
 
-- **scss:** adds new .has-focus class to MyComponent ([660b64b](https://github.com/richmccartney/mono-test/commit/660b64b62780e5b8f1feac6b7ae96838a7f757bc))
+- **scss:** adds new .has-focus class to MyComponent
+  ([660b64b](https://github.com/richmccartney/mono-test/commit/660b64b62780e5b8f1feac6b7ae96838a7f757bc))